refactor: Move `UnaryOp` and `BinaryOp` to `mod op`
