+++ src/vm/mod.rs
@@ -1,20 +1,16 @@
 //! Tokay virtual machine
 mod accept;
-mod binaryop;
 mod capture;
 mod context;
 mod op;
 mod program;
 mod reject;
 mod thread;
-mod unaryop;
 
 pub use accept::*;
-pub use binaryop::*;
 pub use capture::*;
 pub use context::*;
 pub(crate) use op::*;
 pub use program::*;
 pub use reject::*;
 pub use thread::*;
-pub use unaryop::*;
